Module mc_settlement::ethereum::client::starknet_contract

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

ConfigHashCall

Container type for all input parameters for the configHash function with signature configHash() and selector 0xe1f1176d

ConfigHashReturn

Container type for all return fields from the configHash function with signature configHash() and selector 0xe1f1176d

ProgramHashCall

Container type for all input parameters for the programHash function with signature programHash() and selector 0x8a9bf090

ProgramHashReturn

Container type for all return fields from the programHash function with signature programHash() and selector 0x8a9bf090

StarknetContract

StateBlockNumberCall

Container type for all input parameters for the stateBlockNumber function with signature stateBlockNumber() and selector 0x35befa5d

StateBlockNumberReturn

Container type for all return fields from the stateBlockNumber function with signature stateBlockNumber() and selector 0x35befa5d

StateRootCall

Container type for all input parameters for the stateRoot function with signature stateRoot() and selector 0x9588eca2

StateRootReturn

Container type for all return fields from the stateRoot function with signature stateRoot() and selector 0x9588eca2

UpdateStateCall

Container type for all input parameters for the updateState function with signature updateState(uint256[]) and selector 0x853e2461

Enums

StarknetContractCalls

Container type for all of the contract’s call

Statics

STARKNETCONTRACT_ABI

The parsed human-readable ABI of the contract.